Remove lint exception repeated-isinstance-calls
repeated-isinstance-calls (PLR1701) was removed as rule in ruff

Relaxing the python version (issue #6) for the CI made it
use a later python version (3.12.4), which used a later version
of validate-pyproject, which did not recognize the exception
we had placed on PLR1701. The easiest fix was to remove the
exception and satisfy the rule, so that we don't have to
worry about which version of validate-pyproject is used
in the future.