Make session loader look up tranducer
Issue #57

Previously it was wrongly trying to build a Transducer using the
transducer dict inside the session, rather than looking up
the transducer by the transducer ID provided in that dict.

In order to achieve the lookup, the Session needs to know about the
Database. This results in a circular interdependency between Session
and Database. There are multiple solutions, but I went with just
accepting the interdependency since it seems natural here.