Port over solution creation and scaling code (#126)
This adds the `calc_solution` member from `Protocol` which checks the validity of session targets, setup the simulation scene, beamform the signal, run the simulation, analyze the solution and rescale it.

This currently does not perform any segmentation but create a default water volume (with unform isotropic medium), there is an initial implementation for MRI tissue segmentation that will be re-worked in the future.