Standardize example database formatting
This was done by using the Database class to read everything and write
it back out, getting everything into a standard formatting. This will
make comparisons easier in the future, and importantly ensures the mannequin
example follows the standard format.

This is setup work leading up to transducer tracking: #65 and #66