Fix focus masking for solution analysis (#141)
* Fix focus masking for solution analysis

Found one of the sources of regression in #141.  Focal masking was both using offset grid incorrectly (it was calculating the matrix centered at the origin in stead of the point because the default center behavior in `mask_focus` got switched between matlab and python. And the aspect ratio was being applied backwards in `mask_focus`, as a multiplication instead of a division, causing a pancake-shaped ROI instead of a cigar shaped ROI.

* update masking unit test

* Update offset unit test

* fix offset unit test

* fix unit test again