Pin k-wave-python 0.3.4 to avoid a bug in 0.3.5
When using python 3.9, which we need to be able to do as explained
in Issue #167, the latest version of k-wave-python that can be installed
is 0.3.5, and this has a bug that has been affecting our simulation
runs. The fix to the bug takes place in a later version of
k-wave-python that has dropped python 3.9 support.