Install mac dependencies in workflow runner
To continue supporting python 3.9 (issue #167) we had to
pin the version of k-wave-python to one that precedes the introduction
of a certain bug that affects k-wave simulations. For robustness, we
have added a unit test that runs a simulation, but this requires
certain dependencies to be installed for the macos test runners.