Add photocollection concept to database (#234)
Adds a concept of photocollections to the database. Photocollections are
associated with the specific session_id of a specific subject_id, and
they are identified by the reference_number. 'reference_number' was used
instead of photocollection_id for consistency with 3D Open Water. 3D
Open Water takes a collection of `.jpeg` photos and the goal of
photocollections are to store that collection of photos into the
database for downstream processing; specifically, the plan is for a
photocollection of 70-80 photos of a patient's face to be used to create
a mesh (i.e. a "photoscan" object). There is no special object
associated with a photocollection -- it's just a collection of photos
stored in a directory and identified by the reference_number.

- Updates the database to valid state. The updated database includes two
  example photocollections for example_subject