Move ArrayTransform into geo
This is part of work towards #166.
The standoff_transform that goes into virtual fit ought to have
units associated with it, since it may often be provided in transducer
native units which can be different from the volume RAS space units.
So at this point ArrayTransform is more generally useful and doesn't
belong in session.py.