Add annotations to every single dataclass (#262)
This commit adds Annotations[type, OpenLIFUFieldData()] to all data members
of all classes marked with the decorator @dataclass. In places where
docstrings were present, the docstrings were used to populate the
description of the OpenLIFUFieldData. Places where an educated guess
of the meaning of a field could not be made were marked as TODO.