Add failing tests for SolutionAnalysis and SolutionAnalysisOptions (#296)
Theses tests relating to various serialization and to/from dict calls
should pass. The fact that they fail indicate that there are errors with
the serialization. These errors are probably caused by not recursively
calling to/from dict on ParameterConstraint objects and instead
initializing them from dicts. For SolutionAnalysisOptions, the recursive
call needs to be added to the `from_dict()` method. For
SolutionAnalysis, it needs to overwrite `from_dict()` as defined in the
`DictMixin` from which it inherits `from_dict`.