Fix `from_dict` in SolutionAnalysis and SolutionAnalysisOptions (#296)
The previous implementations of `from_dict` did not recursively call
`ParameterConstraint.from_dict()`. `from_dict` was either the vanilla
version inherited from `DictMixin` or did not complete that step.
Because of this, instantiations of `cls` set the `param_constraints` to
a mappings of parameters to dictionaries *resembling* the
`ParameterConstraint` objects instead of actual `ParameterConstraint`s