Run Meshroom nodes sequentially, add timing, and restrict resize parameters (#313 and #314)
This commit returns Python control after each Meshroom node runs. It adds a return_durations
argument to the run_reconstruction function, which, if set to True, returns a dictionary with
the runtime of each node. Additionally, run_reconstruction now raises a ValueError if the
combination of input_resize_width and selected pipeline would result in a Meshroom CUDA error.