Add kwave inert import utility (#377)
This is a way of importing kwave while skipping its automatic download
of binaries. We need to do this in order to get the binary paths from
kwave to support offline use, or delayed download after a prompt, etc.
It's not pretty, but even if we upgrade to the latest k-wave it should
still work. I will add a unit test that uses this to make sure we catch
any issues while upgrading kwave.